NativeLibrary Fix: keep old naming to avoid future conflict
