+++ src/java/org/apache/cassandra/db/commitlog/CommitLogSegment.java
@@ -43,7 +43,7 @@ import org.apache.cassandra.io.util.FileUtils;
 import org.apache.cassandra.schema.SchemaManager;
 import org.apache.cassandra.schema.TableId;
 import org.apache.cassandra.schema.TableMetadata;
-import org.apache.cassandra.utils.NativeLibrary;
+import org.apache.cassandra.utils.INativeLibrary;
 import org.apache.cassandra.utils.IntegerInterval;
 import org.apache.cassandra.utils.concurrent.OpOrder;
 import org.apache.cassandra.utils.concurrent.WaitQueue;
@@ -174,7 +174,7 @@ public abstract class CommitLogSegment
         try
         {
             channel = FileChannel.open(logFile.toPath(), StandardOpenOption.WRITE, StandardOpenOption.READ, StandardOpenOption.CREATE);
-            fd = NativeLibrary.instance.getfd(channel);
+            fd = INativeLibrary.instance.getfd(channel);
         }
         catch (IOException e)
         {
