+++ src/java/org/apache/cassandra/db/lifecycle/LogReplica.java
@@ -31,7 +31,7 @@ import org.apache.cassandra.io.FSError;
 import org.apache.cassandra.io.FSReadError;
 import org.apache.cassandra.io.util.FileUtils;
 import org.apache.cassandra.utils.FBUtilities;
-import org.apache.cassandra.utils.NativeLibrary;
+import org.apache.cassandra.utils.INativeLibrary;
 
 /**
  * Because a column family may have sstables on different disks and disks can
@@ -54,7 +54,7 @@ final class LogReplica implements AutoCloseable
 
     static LogReplica create(File directory, String fileName)
     {
-        int folderFD = NativeLibrary.instance.tryOpenDirectory(directory);
+        int folderFD = INativeLibrary.instance.tryOpenDirectory(directory);
         if (folderFD == -1 && !FBUtilities.isWindows)
             throw new FSReadError(new IOException(String.format("Invalid folder descriptor trying to create log replica %s", directory.path())), directory.path());
 
@@ -63,7 +63,7 @@ final class LogReplica implements AutoCloseable
 
     static LogReplica open(File file)
     {
-        int folderFD = NativeLibrary.instance.tryOpenDirectory(file.parent());
+        int folderFD = INativeLibrary.instance.tryOpenDirectory(file.parent());
         if (folderFD == -1 && !FBUtilities.isWindows)
             throw new FSReadError(new IOException(String.format("Invalid folder descriptor trying to create log replica %s", file.parent().path())), file.parent().path());
 
@@ -120,7 +120,7 @@ final class LogReplica implements AutoCloseable
         try
         {
             if (directoryDescriptor >= 0)
-                NativeLibrary.instance.trySync(directoryDescriptor);
+                INativeLibrary.instance.trySync(directoryDescriptor);
         }
         catch (FSError e)
         {
@@ -144,7 +144,7 @@ final class LogReplica implements AutoCloseable
     {
         if (directoryDescriptor >= 0)
         {
-            NativeLibrary.instance.tryCloseFD(directoryDescriptor);
+            INativeLibrary.instance.tryCloseFD(directoryDescriptor);
             directoryDescriptor = -1;
         }
     }
