+++ src/java/org/apache/cassandra/io/sstable/SSTableRewriter.java
@@ -25,7 +25,6 @@ import com.google.common.annotations.VisibleForTesting;
 import org.apache.cassandra.cache.InstrumentingCache;
 import org.apache.cassandra.cache.KeyCacheKey;
 import org.apache.cassandra.config.DatabaseDescriptor;
-import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.DecoratedKey;
 import org.apache.cassandra.db.compaction.CompactionRealm;
 import org.apache.cassandra.io.sstable.format.RowIndexEntry;
@@ -34,7 +33,7 @@ import org.apache.cassandra.db.lifecycle.ILifecycleTransaction;
 import org.apache.cassandra.db.rows.UnfilteredRowIterator;
 import org.apache.cassandra.io.sstable.format.SSTableReader;
 import org.apache.cassandra.io.sstable.format.SSTableWriter;
-import org.apache.cassandra.utils.NativeLibrary;
+import org.apache.cassandra.utils.INativeLibrary;
 import org.apache.cassandra.utils.concurrent.Transactional;
 
 /**
@@ -172,7 +171,7 @@ public class SSTableRewriter extends Transactional.AbstractTransactional impleme
                 for (SSTableReader reader : transaction.originals())
                 {
                     RowIndexEntry index = reader.getPosition(key, SSTableReader.Operator.GE);
-                    NativeLibrary.instance.trySkipCache(reader.descriptor.fileFor(Component.DATA), 0, index == null ? 0 : index.position);
+                    INativeLibrary.instance.trySkipCache(reader.descriptor.fileFor(Component.DATA), 0, index == null ? 0 : index.position);
                 }
             }
             else
