+++ src/java/org/apache/cassandra/io/util/ChannelProxy.java
@@ -25,7 +25,7 @@ import java.nio.channels.WritableByteChannel;
 import java.nio.file.StandardOpenOption;
 
 import org.apache.cassandra.io.FSReadError;
-import org.apache.cassandra.utils.NativeLibrary;
+import org.apache.cassandra.utils.INativeLibrary;
 import org.apache.cassandra.utils.concurrent.RefCounted;
 import org.apache.cassandra.utils.concurrent.SharedCloseableImpl;
 
@@ -185,7 +185,7 @@ public final class ChannelProxy extends SharedCloseableImpl
 
     public int getFileDescriptor()
     {
-        return NativeLibrary.instance.getfd(channel);
+        return INativeLibrary.instance.getfd(channel);
     }
 
     @Override
