+++ src/java/org/apache/cassandra/io/util/FileHandle.java
@@ -27,7 +27,7 @@ import org.slf4j.LoggerFactory;
 import org.apache.cassandra.cache.ChunkCache;
 import org.apache.cassandra.io.compress.BufferType;
 import org.apache.cassandra.io.compress.CompressionMetadata;
-import org.apache.cassandra.utils.NativeLibrary;
+import org.apache.cassandra.utils.INativeLibrary;
 import org.apache.cassandra.utils.concurrent.Ref;
 import org.apache.cassandra.utils.concurrent.RefCounted;
 import org.apache.cassandra.utils.concurrent.SharedCloseableImpl;
@@ -165,7 +165,7 @@ public class FileHandle extends SharedCloseableImpl
             else
                 return metadata.chunkFor(before).offset;
         }).orElse(before);
-        NativeLibrary.instance.trySkipCache(channel.getFileDescriptor(), 0, position, path());
+        INativeLibrary.instance.trySkipCache(channel.getFileDescriptor(), 0, position, path());
     }
 
     public Rebufferer instantiateRebufferer()
