+++ src/java/org/apache/cassandra/service/StartupChecks.java
@@ -64,7 +64,7 @@ import org.apache.cassandra.io.util.PathUtils;
 import org.apache.cassandra.schema.TableMetadata;
 import org.apache.cassandra.utils.FBUtilities;
 import org.apache.cassandra.utils.JavaUtils;
-import org.apache.cassandra.utils.NativeLibrary;
+import org.apache.cassandra.utils.INativeLibrary;
 import org.apache.cassandra.utils.SigarLibrary;
 
 import static org.apache.cassandra.config.CassandraRelevantProperties.COM_SUN_MANAGEMENT_JMXREMOTE_PORT;
@@ -282,7 +282,7 @@ public class StartupChecks
         public void execute() throws StartupException
         {
             // Fail-fast if the native library could not be linked.
-            if (!NativeLibrary.instance.isAvailable())
+            if (!INativeLibrary.instance.isAvailable())
                 throw new StartupException(StartupException.ERR_WRONG_MACHINE_STATE, "The native library could not be initialized properly. ");
         }
     };
