+++ src/java/org/apache/cassandra/tools/NodeProbe.java
@@ -98,7 +98,7 @@ import com.google.common.collect.Multimap;
 import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.Uninterruptibles;
 import org.apache.cassandra.tools.nodetool.GetTimeout;
-import org.apache.cassandra.utils.NativeLibrary;
+import org.apache.cassandra.utils.INativeLibrary;
 
 /**
  * JMX client operations for Cassandra.
@@ -1714,7 +1714,7 @@ public class NodeProbe implements AutoCloseable
 
     public long getPid()
     {
-        return NativeLibrary.instance.getProcessID();
+        return INativeLibrary.instance.getProcessID();
     }
 
     public void resumeBootstrap(PrintStream out) throws IOException
