+++ src/java/org/apache/cassandra/utils/HeapUtils.java
@@ -114,7 +114,7 @@ public final class HeapUtils
      */
     public static Long getProcessId()
     {
-        long pid = NativeLibrary.instance.getProcessID();
+        long pid = INativeLibrary.instance.getProcessID();
         if (pid >= 0)
             return pid;
