+++ src/java/org/apache/cassandra/utils/UUIDGen.java
@@ -400,7 +400,7 @@ public class UUIDGen
         }
 
         // Identify the process on the load: we use both the PID and class loader hash.
-        long pid = NativeLibrary.instance.getProcessID();
+        long pid = INativeLibrary.instance.getProcessID();
         if (pid < 0)
             pid = new Random(System.currentTimeMillis()).nextLong();
         updateWithLong(hasher, pid);
